requirements.txt
Django==3.0
gunicorn==20.0.4
djangorestframework
twilio